+++ proto/vote/exported/v1beta1/types.proto
@@ -1,7 +1,7 @@
 syntax = "proto3";
 package vote.exported.v1beta1;
 
-option go_package = "github.com/axelarnetwork/axelar-core/x/vote/exported2";
+option go_package = "github.com/axelarnetwork/axelar-core/x/vote/exported-0.17";
 
 import "gogoproto/gogo.proto";
 import "google/protobuf/any.proto";
