+++ proto/vote/v1beta1/types.proto
@@ -1,7 +1,7 @@
 syntax = "proto3";
 package vote.v1beta1;
 
-option go_package = "github.com/axelarnetwork/axelar-core/x/vote/types2";
+option go_package = "github.com/axelarnetwork/axelar-core/x/vote/types-0.17";
 
 import "google/protobuf/any.proto";
 import "gogoproto/gogo.proto";
