+++ x/vote/keeper/migrate.go
@@ -14,11 +14,139 @@ import (
 // - migrate all completed polls
 func GetMigrationHandler(k Keeper) func(ctx sdk.Context) error {
 	return func(ctx sdk.Context) error {
+		err:= migrateVotes(ctx, k)
+		if err!= nil{
+			return err
+		}
+
 		deleteAllPendingPolls(ctx, k)
 		migrateAllCompletedPolls(ctx, k)
 
 		return nil
 	}
+}
+
+
+func migrateVotes(ctx sdk.Context, k Keeper) error {
+	metadatas := k.getPollMetadatasOld(ctx)
+	for _, metadata := range metadatas {
+		newMetadata := exported.PollMetadata{
+			Key:              metadata.Key,
+			ExpiresAt:        metadata.ExpiresAt,
+			Result:           MigrateVoteData(k.cdc, metadata.RewardPoolName, metadata.Result, k.Logger(ctx)),
+			VotingThreshold:  metadata.VotingThreshold,
+			State:            metadata.State,
+			MinVoterCount:    metadata.MinVoterCount,
+			Voters:           metadata.Voters,
+			TotalVotingPower: metadata.TotalVotingPower,
+			RewardPoolName:   metadata.RewardPoolName,
+		}
+
+		pollStore := k.newPollStore(ctx, newMetadata.Key)
+		pollStore.SetMetadata(newMetadata)
+		votes := pollStore.GetVotesOld()
+		for _, vote := range votes {
+			newVote := types.TalliedVote{
+				Tally:  vote.Tally,
+				Voters: vote.Voters,
+				Data:   MigrateVoteData(k.cdc, metadata.RewardPoolName, vote.Data, k.Logger(ctx)),
+			}
+			pollStore.Set(votesPrefix.AppendStr(pollStore.key.String()).AppendStr(newVote.Hash()), &newVote)
+		}
+	}
+
+	return assertMigrationSuccessful(ctx, k)
+}
+
+func assertMigrationSuccessful(ctx sdk.Context, k Keeper) error {
+	metadatas := k.getPollMetadatas(ctx)
+	for _, metadata := range metadatas {
+		if metadata.Result.GetCachedValue() == nil {
+			return fmt.Errorf("failed to verify poll result for %s", hex.EncodeToString(k.cdc.MustMarshalLengthPrefixed(&metadata)))
+		}
+
+		pollStore := k.newPollStore(ctx, metadata.Key)
+		votes := pollStore.GetVotes()
+		for _, vote := range votes {
+			if metadata.Result.GetCachedValue() == nil {
+				return fmt.Errorf("failed to verify tallied vote data for %s", hex.EncodeToString(k.cdc.MustMarshalLengthPrefixed(&vote)))
+			}
+		}
+	}
+	return nil
+}
+
+// MigrateVoteData migrates vote results from an Any slice to a single Any value
+func MigrateVoteData(cdc codec.BinaryCodec, chain string, data *codectypes.Any, logger log.Logger) *codectypes.Any {
+	switch d := data.GetCachedValue().(type) {
+	case *exported0_17.Vote:
+		events, err := unpackEvents(cdc, d.Results)
+		if err != nil {
+			logger.Error("failed tp unpack vote results for %s", hex.EncodeToString(cdc.MustMarshalLengthPrefixed(d)))
+			return nil
+		}
+
+		if len(events) != 0 {
+			chain = events[0].Chain
+		}
+		result, err := evmtypes.PackEvents(chain, events)
+		if err != nil {
+			logger.Error("failed to pack events for %s", hex.EncodeToString(cdc.MustMarshalLengthPrefixed(d)))
+			return nil
+		}
+		packedVote, err := codectypes.NewAnyWithValue(&exported.Vote{Result: result})
+		if err != nil {
+			logger.Error("failed to pack vote for %s", hex.EncodeToString(cdc.MustMarshalLengthPrefixed(d)))
+			return nil
+		}
+		return packedVote
+	default:
+		// in all other cases the data struct stays the same
+		logger.Info(fmt.Sprintf("ignoring migration for data of type %s", data.GoString()))
+		return data
+	}
+}
+
+func (k Keeper) getPollMetadatasOld(ctx sdk.Context) []exported0_17.PollMetadata {
+	var pollMetadatas []exported0_17.PollMetadata
+
+	iter := k.getKVStore(ctx).Iterator(pollPrefix)
+	utils.CloseLogError(iter, k.Logger(ctx))
+
+	for ; iter.Valid(); iter.Next() {
+		var pollMetadata exported0_17.PollMetadata
+		k.cdc.MustUnmarshalLengthPrefixed(iter.Value(), &pollMetadata)
+		pollMetadatas = append(pollMetadatas, pollMetadata)
+	}
+
+	return pollMetadatas
+}
+
+func (p *pollStore) GetVotesOld() []types0_17.TalliedVote {
+	iter := p.Iterator(votesPrefix.AppendStr(p.key.String()))
+	defer utils.CloseLogError(iter, p.logger)
+
+	var votes []types0_17.TalliedVote
+	for ; iter.Valid(); iter.Next() {
+		var vote types0_17.TalliedVote
+		iter.UnmarshalValue(&vote)
+		votes = append(votes, vote)
+	}
+	return votes
+}
+
+// UnpackEvents converts Any slice to Events
+func unpackEvents(cdc codec.BinaryCodec, eventsAny []*codectypes.Any) ([]evmtypes.Event, error) {
+	var events []evmtypes.Event
+	for _, e := range eventsAny {
+		var event evmtypes.Event
+		if err := cdc.Unmarshal(e.Value, &event); err != nil {
+			return nil, err
+		}
+		events = append(events, event)
+	}
+
+	return events, nil
 }
 
 func deleteAllPendingPolls(ctx sdk.Context, k Keeper) {
